Fix incorrect counting in for packets related to standby interface in dualtor (#43)
Why I do it?
In dualtor, we have below 2 scenarios, which requires counter to be align with dhcp relay

DHCP packets come from standby downstream interfaces would be dropped, hence we don't need to count them
DHCP packets to standby downstream interfaces would be relayed, hence we need to count them
Add support to per-interface counter for above 2 scenarios
How I did it?
When a RX packet come, it would go through Physical interface and context interface, which means it would be captured once on the context interface and once on the physical interface.
1.1 In non-dualtor and RX from uplink in dualtor, we can directly increase corresponding counter when captured it in physical interface and context interface because they are not related to standby interface.
1.2 In RX from downlink in dualtor, since we cannot know whether a packet come from standby interface or not when captured it in context interface, hence we cannot increase counter here. We need to increase both physical interface counter and context interface counter when captured it in physical interface based on whether it's standby interface

When a TX packet out, it would go through context interface and physical interface, we don't need to care about whether it's standby interface or not, we can directly increase corresponding counter when captured it in context interface and physical interface

How I verify it?
Manually test
Run test_dhcp_relay in dualtor and t0