Update clear_counter_timeout to fix clear counter issue (#48)
### Why I did it
When we clear dhcp_relay ipv4 counter, cli would send a SIGUSR1 to stop writing DB, then sends a SIGUSR2 to sync DB to cache and continue timer to write DB.
If writing DB happens during SIGUSR1 and SIGUSR2, it would be ignored. But if the SIGUSR2 is not received after SIGUSR1 for a while (5s after last writing DB), it would timeout to break clear counter procedure.
There is such [fail to clear counter] with default startup parameter (Interval to write DB set to 20s):
1) 1st writing DB happened in 00:00:00.00
2) Clear CLI sent SIGUSR1 in 00:00:19.95
3) 2nd writing DB happened in 00:00:20.00 (Because there is already 20s since last update, which > 5s, it would interrupt clear procedure)
4) Clear CLI sent SIGUSR2 in 00:00:20.20 but it fails

### How I did it
If the db_update_interval < 5, then the clear timeout is set to 5, else, the clear timeout would be set to [db_update_interval + 1], it can make sure procedure of clearing counter wouldn't be interrupted by writing DB

### How I verify it
Manually test